Change du_dp outputs to ull in potential interface

+++ timemachine/cpp/src/periodic_torsion.hpp
@@ -32,7 +32,7 @@ public:
         const double *d_box,
         const double lambda,
         unsigned long long *d_du_dx,
-        double *d_du_dp,
+        unsigned long long *d_du_dp,
         unsigned long long *d_du_dl,
         unsigned long long *d_u,
         cudaStream_t stream) override;
